+++ pkg/workloads/cortex/lib/storage/s3.py
@@ -251,9 +251,7 @@ class S3(object):
         return json.loads(obj.decode("utf-8"))
 
     def list_objects(self, path):
-        bucket, prefix = self.deconstruct_s3_path(path)
+        _, prefix = self.deconstruct_s3_path(path)
         return [
-            obj["Key"]
-            for obj in self.s3.list_objects(Bucket=bucket, Prefix=prefix)["Contents"]
-            if obj["Key"][:1] != "/"
+            obj[len(prefix) + 1 :] for obj in self.search(prefix=prefix) if not obj.endswith("/")
         ]
