+++ pkg/workloads/cortex/tf_api/api.py
@@ -31,6 +31,7 @@ from google.protobuf import json_format
 from cortex import consts
 from cortex.lib import util, tf_lib, package, Context
 from cortex.lib.log import get_logger
+from cortex.lib.storage import S3, LocalStorage
 from cortex.lib.exceptions import CortexException, UserRuntimeException, UserException
 from cortex.lib.context import create_transformer_inputs_from_map
 
@@ -445,15 +446,16 @@ def get_signature(app_name, api_name):
 def download_dir_external(ctx, s3_path, local_path):
     util.mkdir_p(local_path)
     bucket_name, prefix = ctx.storage.deconstruct_s3_path(s3_path)
-    objects = ctx.storage.list_objects(s3_path)
+    storage_client = S3(bucket_name, client_config={})
+    objects = storage_client.list_objects(s3_path)
     version = prefix.split("/")[-1]
+    local_path = os.path.join(local_path, version)
     for obj in objects:
-        local_key = obj[len(prefix) - len(version) :]
-        if not os.path.exists(os.path.dirname(local_key)):
-            util.mkdir_p(os.path.join(local_path, os.path.dirname(local_key)))
+        if not os.path.exists(os.path.dirname(obj)):
+            util.mkdir_p(os.path.join(local_path, os.path.dirname(obj)))
 
         ctx.storage.download_file_external(
-            bucket_name + "/" + obj, os.path.join(local_path, local_key)
+            bucket_name + "/" + os.path.join(prefix, obj), os.path.join(local_path, obj)
         )
 
 
@@ -497,10 +499,13 @@ def start(args):
 
         if not os.path.isdir(args.model_dir):
             if util.is_resource_ref(api["model"]):
+                model_name = util.get_resource_ref(api["model"])
+                model = ctx.models[model_name]
                 ctx.storage.download_and_unzip(model["key"], args.model_dir)
             else:
                 download_dir_external(ctx, api["model"], args.model_dir)
-
+        local = LocalStorage(base_dir=args.model_dir)
+        util.logger.info(local.list_objects(args.model_dir))
         if args.only_download:
             return
